removed unnecessary steps
