+++ src/components/TourButton/TourButton.module.css
@@ -1,4 +1,4 @@
-/* .tourButton {
+.tourButton {
     position: fixed;
     bottom: 10px;
     right: 5px;
@@ -8,10 +8,43 @@
     z-index: 1000;
     background-color: transparent;
     border-style: none;
+    padding: 0;
+}
+
+.tourButton:focus {
+    outline: none;
+}
+
+.tourButton:focus-visible {
+    outline: none;
 }
 
 .question {
     width: 40px;
     height: 40px;
     fill: #555;
-} */
+}
+
+.shake {
+    animation: shake 1s ease-in-out;
+}
+
+@keyframes shake {
+
+    0%,
+    100% {
+        transform: translateX(0);
+    }
+
+    25% {
+        transform: translateX(-5px);
+    }
+
+    50% {
+        transform: translateX(5px);
+    }
+
+    75% {
+        transform: translateX(-5px);
+    }
+}
