+++ src/pages/TrackerPage/TrackerPage.jsx
@@ -1,55 +1,12 @@
-import React, { useEffect } from "react";
+import React from "react";
 import { Helmet } from "react-helmet-async";
 import { Page } from "../../components/Page/Page";
 import WaterMainInfo from "../../components/WaterMainInfo/WaterMainInfo";
 import WaterDetailedInfo from "../../components/WaterDetailedInfo/WaterDetailedInfo";
 import css from "./TrackerPage.module.css";
-// import TourButton from "../../components/TourButton/TourButton";
-import iconsPath from "../../assets/icons/sprite.svg";
-import steps from "../../helpers/steps";
-import { useTour } from '@reactour/tour';
+import TourButton from "../../components/TourButton/TourButton";
 
 export default function TrackerPage() {
-
-  const { setCurrentStep, setIsOpen, isOpen } = useTour();
-
-  useEffect(() => {
-  if (isOpen) {
-      document.body.style.overflow = "hidden";
-  } else {
-      document.body.style.overflow = "unset";
-  }
-  }, [isOpen]);
-
-    const startTour = () => {
-      setCurrentStep(0);
-      const firstStepSelector = steps[0].selector;
-      const waitForSteps = setInterval(() => {
-      const stepElement = document.querySelector(firstStepSelector);
-      if (stepElement) {
-      clearInterval(waitForSteps);
-      setIsOpen(true);
-      }
-  }, 200);
-  };
-
-
-  useEffect(() => {
-    const questionIcon = document.querySelector(`.${css.question}`);
-
-    const addShakeAnimation = () => {
-      if (questionIcon) {
-        questionIcon.classList.add(css.shake);
-        setTimeout(() => {
-          questionIcon.classList.remove(css.shake);
-        }, 1000); // Довжина анімації (1 секунда)
-      }
-    };
-
-    const intervalId = setInterval(addShakeAnimation, 6000); // Запуск кожну хвилину
-
-    return () => clearInterval(intervalId); // Очищення інтервалу при розмонтуванні
-  }, []);
 
   return (
     <React.Fragment>
@@ -60,15 +17,7 @@ export default function TrackerPage() {
         <div className={css.trackerPage}>
           <WaterMainInfo />
           <WaterDetailedInfo />
-            <button
-              data-tour="step-start"
-              className={css.tourButton}
-              onClick={startTour}>
-                <svg className={css.question} width="20" height="20">
-                    <use href={`${iconsPath}#icon-question`} />
-                </svg>
-            </button>
-            {/* <TourButton/> */}
+            <TourButton/>
         </div>
       </Page>
     </React.Fragment>
